Refactor render directory
