+++ src/index.js
@@ -1,14 +1,14 @@
 const Validator = require('./validator');
 const usfmHelps = require('./usfmHelps');
 const usfmJsHelps = require('./usfmJsHelps');
-const PipelineHandler = require('./classes/PipelineHandler');
-const ProskommaRender = require('./ProskommaRender');
-const PerfRenderFromJson = require('./PerfRenderFromJson');
-const PerfRenderFromProskomma = require('./PerfRenderFromProskomma');
-const SofriaRenderFromJson = require('./SofriaRenderFromJson');
-const SofriaRenderFromProskomma = require('./SofriaRenderFromProskomma');
-const mergeActions = require('./mergeActions');
-const pipelines = require('./pipelines');
+const PipelineHandler = require('./render/classes/PipelineHandler');
+const ProskommaRender = require('./render/renderers/ProskommaRender');
+const PerfRenderFromJson = require('./render/renderers/PerfRenderFromJson');
+const PerfRenderFromProskomma = require('./render/renderers/PerfRenderFromProskomma');
+const SofriaRenderFromJson = require('./render/renderers/SofriaRenderFromJson');
+const SofriaRenderFromProskomma = require('./render/renderers/SofriaRenderFromProskomma');
+const mergeActions = require('./render/renderers/mergeActions');
+const pipelines = require('./render/pipelines');
 const render = require('./render');
 
 module.exports = {
