+++ src/render/pipelines/index.js
