+++ src/render/pipelines/justTheBiblePipeline.json
