+++ src/render/pipelines/mergeAlignmentPipeline.json
