+++ src/render/pipelines/stripAlignmentPipeline.json
