+++ src/render/pipelines/stripUwAlignmentPipeline.json
