+++ src/render/renderers/PerfRenderFromJson.js
