+++ src/render/renderers/PerfRenderFromProskomma.js
