+++ src/render/renderers/SofriaRenderFromJson.js
