+++ src/render/renderers/SofriaRenderFromProskomma.js
