+++ src/render/renderers/mergeActions.js
